Add Raster class with some basic operations. Needs fleshing out
Add the base Raster class. This class will be the backbone of the project. This is very bare bones and a little hacky. It can read in tiff files but cannot save to tiff because of limitations in xarray. This needs to be investigated further. I'm not convinced that xarray/rasterio is the way forward because there are
a lot of limitations when handling tiffs