Turn off tiff tile-size-aware chunking until we figure out what causes the dask warning
The following warning pops up when chunknig the data:
xarray/core/indexing.py:1229: PerformanceWarning: Slicing is producing a large chunk. To accept the large
chunk and silence this warning, set the option
    >>> with dask.config.set(**{'array.slicing.split_large_chunks': False}):
    ...     array[indexer]

To avoid creating the large chunks, set the option
    >>> with dask.config.set(**{'array.slicing.split_large_chunks': True}):
    ...     array[indexer]
  return self.array[key]

It is unclear if the data is being loaded so I'm avoiding the suspicious operation until I know more.