Add workaround to fix invalid build results due to a bug in pip
Pip installs the deps as specified in the requirements file. When the
cython modules are built, however, pip installs the latest version of
numpy, ignoring the specified version, and then builds against it.
This creates invalid shared objects that can't be used. This would not
be an issue normally but for a confluance of circumstances that expose
the bug in pip. These are: Numpy recently changed the C API and
Numba is incompatable with numpy > 1.21. Below are reference links.

https://github.com/pypa/pip/issues/9542
https://github.com/scikit-learn-contrib/hdbscan/issues/457#issuecomment-773671043
https://stackoverflow.com/questions/66060487/valueerror-numpy-ndarray-size-changed-may-indicate-binary-incompatibility-exp