Refactor vector.py to make a handle data length
Dask doesn't keep track of a dataframe's lenght. len(df) causes
computation. df.divisions is a tuple of partition division points and is
the best way to estimate the length without loading the data.
Unfortunatly dask special cases the last partition. The last partitions
end division point is inclusive while the rest are exclusive. Dask
doesn't differentiate the last partition in any way, however. It can
be shuffled or dropped so it is impossible to be sure what partittion is
actually the original last. This makes getting the exact length
imposible without loading data. An estimate of the length based on
divisions is prone to be off by one.

This commit refactors the module to make a best effort estimate of the
actual size.