Refactor to allow rasters to be used as features
Delay application of labeled_comprehension

labeled_comprehension is lazy however for a large enough input, the
dask overhead in the iteration adds up to a significant delay. The final
result size is known so the application of labeled_comprehension can be
delayed. This commit wraps it in a delayed object.