Fix test borken by upstream change
Looks like a change to dask has made it so that dropping the geometry
column from a dask_goepandas dataframe causes the dataframe to change
types to a dask dataframe. I'm not sure what change caused this but it
must be in dask rather than dask_goepandas. My fix is to simply drop the
type check since a dask dataframe works just as well for the table
property.