Revert to decreasing y dim to follow convention.
This updates raster_tools to follow the common convention of having the
y dim decrease with increasing row numbers. This produces a negative
resolution for the y dim but make other things easier such as plotting
and comparison with ESRI rasters. It also saves a step when loading in
rasters since the flip is no longer needed.