Refactor how null value handling to force dtype consistency
The null value will now be consistent with the raster dtype. This is done because of rioxarray's null value handling. rioxarray's write_nodata method fails if the null value does not match the dtype. Unfortunately, that method is the only way to make rioxarray aware of the null value, which is necessary for preserving the null value to disk when saving.