Fix lines_stats.length issues
This drops pygeos in favor of shapely 2.0 for the geopandas backend and
uses batching to better control memory usage.

pygeos randomly converts LineStrings to LinearRings which then causes it
to throw an error during bounds checking. This seems to be random and
could not be replicated on the main thread. This commit switches to
shapely 2.0 which is much more stable and seems to be slightly faster.

This commit also breaks dataframes chunks into batches when computing
the length stats. This was done because the buffering and rasterizing of
the geometries causes large swings in memory usage and can easily fill
system memory when enough dask workers are running.