Refactor and optimize rasterization
This refactors the rasterization machinery to take advantage of spatial partitions when available to reduce the number of chunk operations. It moves the rasterization code out of vector.py and into rasterize.py. It replaces the old tests for rasterization with new ones. It also adds 2 support methods to the raster class.