Refactor zonal_stats to use dataframe based aggregation.
This refactors the zonal_stat function to use dataframe groupby aggregation instead of dask_image's labeled_comprehension for the zonal stat application. It is faster, less error prone, and easier to understand. It also doesn't require knowing the zone ID values ahead of time.

This also updates the output dataframe to be in wide format by default with the zone IDs in the index. It adds the `wide_format` option to give the user control over long/wide. An example of `wide_format=True` is below:

```
Dask DataFrame Structure:
                band_1            band_2            band_3
                  mean      min     mean      min     mean      min
npartitions=1
               float64  float64  float64  float64  float64  float64
                   ...      ...      ...      ...      ...      ...
```

and for `wide_format=False`:

```
Dask DataFrame Structure:
                zone   band     mean      min
npartitions=1
               int64  int64  float64  float64
                 ...    ...      ...      ...
```

Finally, this update changes the name of the stat function to count the number of unique values from 'unique' to 'nunique'.