Handle dask corner case where the computed result is not writeable
If the result of `dask.array.full_like` has only a single chunk, the
computed numpy array will not be writeable. This causes issues in
raster_tools.creation.{full,zeros,ones}_like as well as when creating an
unmasked raster. This commit handles the corner case and adds tests to
catch it.

 ref: https://github.com/dask/dask/issues/11531