Add workaround to avoid flakey Dask behavior
This commit adds a workaround that avoids failures caused by dask
computing an invalid results. This would happend randomly and very
rarley. The issue is that when a dask-dataframe partition is passed to a
map_blocks call as one of the args, it will sometimes evaluate to a
pandas.Series object with a single element containing a dask graph
key-tuple(?). Something like:

```
0    (getitem-to_frame-b2fa15abf...", 0)
dtype: object
```

I believe the issue originates in the dask-expr backend but I don't know
how to debug it. This commit works around that by converting dataframe
paritions to delayed objects before passing them to map_blocks.